Add files via upload
The first part of the code involves importing necessary libraries and loading the "Review" file containing various reviews. After that, the ratings, ranging from 1 to 5, are adjusted to ensure each class has a total of 29770 instances, achieving balanced classes. Subsequently, the comments are categorized into positive, negative, and neutral sentiments, and the data is split before preprocessing is applied.